+++ packages/runner/injection/patches/cross-origin/fetch.ts
