+++ packages/runner/injection/patches/cross-origin/utils/index.ts
